@@ -1,9 +1,9 @@
 module github.com/MRtecno98/bucket
 
-go 1.18
+go 1.19
 
 require (
-	github.com/MRtecno98/afero v1.9.3-0.20220615121724-57f06c0516c5
+	github.com/MRtecno98/afero v1.9.3
 	github.com/urfave/cli/v2 v2.8.1
 	golang.org/x/exp v0.0.0-20220613132600-b0d781184e0d
 	gopkg.in/yaml.v3 v3.0.1
@@ -15,10 +15,11 @@ require (
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/kr/fs v0.1.0 // indirect
+	github.com/kr/pretty v0.2.0 // indirect
 	github.com/pkg/sftp v1.13.5 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
-	github.com/spf13/afero v1.8.2 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
 	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e // indirect
 	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
